% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsmglobal-package.R
\docType{package}
\name{fsmglobal-package}
\alias{fsmglobal-package}
\title{fsmglobal: Global Faecal Sludge Emptying Services Demand}
\description{
Global faecal sludge emptying demand data for 175 countries and territories, based on 2017 population figures. Demand is quantified by the population served by one of four emptying methods (mechanized, non-mechanized, unemptiable, no facility). The demand is disaggregated into urban and rural proportions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/openwashdata/fsmglobal}
  \item Report bugs at \url{https://github.com/openwashdata/fsmglobal/issues}
}

}
\author{
\strong{Maintainer}: Lars Schöbitz \email{lschoebitz@ethz.ch} (\href{https://orcid.org/0000-0003-2196-5015}{ORCID})

Authors:
\itemize{
  \item Nicola Greene \email{nicola@opero-services.com} (\href{https://orcid.org/0000-0001-9480-4197}{ORCID})
  \item Sarah Hennessy \email{sarah@triangle-environmental.com} (\href{https://orcid.org/0009-0000-4711-0766}{ORCID})
  \item Tate W. Rogers \email{tate@triangle-environmental.com} (\href{https://orcid.org/0009-0009-9006-5994}{ORCID})
  \item Jocelyn Tsai \email{jct33@unc.edu} (\href{https://orcid.org/0000-0001-8777-0339}{ORCID})
  \item Francis L. de los Reyes III \email{fldelosr@ncsu.edu} (\href{https://orcid.org/0000-0002-3593-0932}{ORCID})
}

}
\keyword{internal}
