% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsmglobal-package.R
\docType{package}
\name{fsmglobal-package}
\alias{fsmglobal-package}
\alias{_PACKAGE}
\title{fsmglobal: Global Faecal Sludge Emptying Services Demand}
\description{
Global faecal sludge emptying demand data for 175 countries. Demand is quantified by population size receiving one of four emptying methods (mechanized, non-mechanized, unemptiable, no facility). The demand is disaggregated by rural and urban percentages of population.
}
\keyword{internal}
